RAG_visual_lab/pyproject.toml
[project]
name = "rag-visual-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
   "chromadb>=1.1.0",
   "crewai>=0.201.1",
   "google-genai>=1.38.0",
   "markitdown[all]>=0.1.3",
   "openai>=1.109.0",
   "python-dotenv>=1.1.1",
   "sentence-transformers>=5.1.1",
   "streamlit>=1.50.0",
   "streamlit[pdf]",
   "PyPDF2==3.0.1",
   "streamlit-elements==0.1.*",
   "lightrag-hku==1.4.9.4",
   "plotly==6.3.1",
   "streamlit-agraph>=0.0.45",
   "pypdf==6.1.3",
   "python-docx==1.2.0",
   "redis>=6.4.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.10",
    "fakeredis==2.32.0",
]

